pyproject.toml
[tool.poetry]
package-mode = false
name = "steamforlinux"
version = "0.1.0"
description = ""
authors = ["AlejandroFeriaGonzalez <aferiag@unal.edu.co>"]
readme = "README.md"
packages = [{include = "steamforlinux", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
beautifulsoup4 = "^4.12.3"
pydantic = "^2.8.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
